requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.49.0
dataclasses-json==0.6.7
datasets==3.5.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecated==1.2.18
dill==0.3.8
distro==1.9.0
docker==7.1.0
durationpy==0.9
evaluate==0.4.3
executing==2.2.0
fastapi==0.115.9
fastjsonschema==2.21.1
filelock==3.18.0
flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
ipykernel==6.29.5
ipython==9.2.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-ollama==0.3.0
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langfuse==2.60.3
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.24
litellm==1.65.5
lxml==5.4.0
mako==1.3.9
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mlflow==2.21.3
mlflow-skinny==2.21.3
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.2
notebook-shim==0.2.4
numpy==2.2.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.21.0
openai==1.72.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.8
posthog==3.23.0
prometheus-client==0.22.0
prompt-toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pure-eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pygments==2.19.1
pyparsing==3.2.3
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
send2trash==1.8.3
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.3
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.0
typer==0.15.2
types-python-dateutil==2.9.0.20250516
typing-extensions==4.13.1
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
waitress==3.0.2
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
